# src/shopping-admin/shopping-admin-consumer/docs/swagger.yaml
basePath: /api/shopping/admin
definitions:
...
  Request-any:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
  Request-shopping_admin_AdParam:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        allOf:
        - $ref: '#/definitions/shopping_admin.AdParam'
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
  Request-shopping_admin_BrandParam:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        allOf:
        - $ref: '#/definitions/shopping_admin.BrandParam'
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
  Request-shopping_admin_BusinessFranchiseeParam:
    properties:
      pageIndex:
        description: 分页索引
        example: 1
        minimum: 1
        type: integer
      pageSize:
        description: 分页长度
        example: 10
        maximum: 100
        minimum: 10
        type: integer
      params:
        allOf:
        - $ref: '#/definitions/shopping_admin.BusinessFranchiseeParam'
        description: 请求查询参数
    required:
    - pageIndex
    - pageSize
    type: object
...
  Response:
    properties:
      code:
        default: 0
        description: 结果状态码
        type: integer
      data:
        description: 结果
      msg:
        default: 请求失败
        description: 状态信息
        type: string
    type: object
  shopping_admin.Ad:
    properties:
      adId:
        description: '@gotags:  gorm:"primary_key;<-:create" validate:"required"'
        type: string
      adInfo:
        description: 广告信息json数组：[{"title":"商品标题","url":"广告图地址","goodId":"跳转的商品id"}]
        items:
          $ref: '#/definitions/shopping_admin.AdInfo'
        type: array
      adType:
        description: 广告类型(1:全网推广广告，2:商家自定义广告)
        type: integer
      position:
        description: 广告位置
        type: integer
      storeId:
        description: 店铺id
        type: string
      viewType:
        description: 呈现类型(1:轮播图广告，2:单图片广告)
        type: integer
    required:
    - adId
    - adType
    - position
    - storeId
    - viewType
    type: object
  shopping_admin.AdInfo:
    properties:
      imageUrl:
        description: 广告图片地址
        type: string
      targetUrl:
        description: 跳转地址
        type: string
      title:
        description: 广告标题
        type: string
    type: object
  shopping_admin.AdParam:
    type: object
  shopping_admin.Brand:
    properties:
      brandId:
        description: 品牌id
        type: string
      displayType:
        description: 展示类型（1：图片，0：文字）
        type: integer
      introduction:
        description: 介绍
        type: string
      logo:
        description: logo
        type: string
      name:
        description: 名称
        type: string
      sortBy:
        description: 排序
        type: integer
      url:
        description: 网址
        type: string
    type: object
  shopping_admin.BrandParam:
    type: object
...
  shopping_admin.BusinessFranchisee:
    type: object
  shopping_admin.BusinessFranchiseeParam:
    type: object
...
info:
  contact:
    email: 826422592@qq.com
    name: 杨大琼
    url: http://localhost
  description: swagger go API文档
  license:
    name: Apache 2.0
  title: 商城后台管理服务提供者接口文档
paths:
  /ad:
    delete:
      description: 主要用于广告删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 广告是删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除广告
      tags:
      - 广告管理
    post:
      description: 主要用于广告新增
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 新增广告参数
        in: body
        name: advertisement
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.Ad'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 创建广告
      tags:
      - 广告管理
    put:
      description: 主要用于广告修改
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 修改广告参数
        in: body
        name: advertisement
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.Ad'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 修改广告
      tags:
      - 广告管理
  /ad/findById/{id}:
    get:
      description: 根据id查询广告内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 广告id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询广告内容详情
      tags:
      - 广告管理
  /ad/findPageList:
    post:
      description: 主要用于查询广告分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-shopping_admin_AdParam'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 广告分页列表
      tags:
      - 广告管理
  /brand:
    delete:
      description: 主要用于品牌删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除品牌
      tags:
      - 品牌管理
    post:
      description: 主要用于品牌新增
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 新增品牌参数
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.Brand'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 创建品牌
      tags:
      - 品牌管理
    put:
      description: 主要用于品牌修改
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 修改品牌参数
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.Brand'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 修改品牌
      tags:
      - 品牌管理
  /brand/findById/{id}:
    get:
      description: 根据id查询品牌内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 品牌id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询品牌内容详情
      tags:
      - 品牌管理
...
  /brand/findPageList:
    post:
      description: 主要用于查询品牌分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-shopping_admin_BrandParam'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 品牌分页列表
      tags:
      - 品牌管理
...
  /businessCircle:
    delete:
      description: 主要用于商圈删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商圈是删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除商圈
      tags:
      - 商圈管理
  /businessCircle/findById/{id}:
    get:
      description: 根据id查询商圈内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商圈id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询商圈内容详情
      tags:
      - 商圈管理
  /businessCircle/findPageList:
    post:
      description: 主要用于查询商圈分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-any'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 商圈分页列表
      tags:
      - 商圈管理
...
  /franchisee:
    delete:
      description: 主要用于商家入住删除
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 删除参数
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 删除商家入住
      tags:
      - 商家入住管理
    post:
      description: 主要用于商家入住新增
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 新增商家入住参数
        in: body
        name: franchisee
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.BusinessFranchisee'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 创建商家入住
      tags:
      - 商家入住管理
    put:
      description: 主要用于商家入住修改
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 修改商家入住参数
        in: body
        name: franchisee
        required: true
        schema:
          $ref: '#/definitions/shopping_admin.BusinessFranchisee'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 修改商家入住
      tags:
      - 商家入住管理
  /franchisee/findById/{id}:
    get:
      description: 根据id查询商家入住内容详情
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 商家入住id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 根据id查询商家入住内容详情
      tags:
      - 商家入住管理
  /franchisee/findPageList:
    post:
      description: 主要用于查询商家入住分页列表
      parameters:
      - default: Bearer token
        description: 授权令牌
        in: header
        name: accessToken
        required: true
        type: string
      - description: 查询参数
        in: body
        name: pageInfo
        required: true
        schema:
          $ref: '#/definitions/Request-shopping_admin_BusinessFranchiseeParam'
      responses:
        "200":
          description: 请求成功
          schema:
            $ref: '#/definitions/Response'
        "400":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
        "500":
          description: 请求异常
          schema:
            $ref: '#/definitions/Response'
      summary: 商家入住分页列表
      tags:
      - 商家入住管理
...
swagger: "2.0"